chore: update core storage

+++ contracts/cosmwasm-vm/cw-ibc-core/src/state.rs
@@ -6,7 +6,7 @@ pub struct CwIbcStore<'a> {
     next_sequence_send: Map<'a, (PortId, ChannelId), Sequence>,
     next_sequence_recv: Map<'a, (PortId, ChannelId), Sequence>,
     next_sequence_ack: Map<'a, (PortId, ChannelId), Sequence>,
-    next_client_sequence: Item<'a, u128>,
+    next_client_sequence: Item<'a, u64>,
     next_connection_sequence: Item<'a, u128>,
     next_channel_sequence: Item<'a, u128>,
     client_connections: Map<'a, ClientId, ConnectionId>,
@@ -71,7 +71,7 @@ impl<'a> CwIbcStore<'a> {
         &self.next_sequence_ack
     }
 
-    pub fn next_client_sequence(&self) -> &Item<'a, u128> {
+    pub fn next_client_sequence(&self) -> &Item<'a, u64> {
         &self.next_client_sequence
     }
     pub fn next_connection_sequence(&self) -> &Item<'a, u128> {
